Updated type sizes and using the newer syscalls

+++ include/kernel/netbsd/sys/time.h
@@ -26,7 +26,7 @@ typedef int suseconds_t;
 # endif
 # ifndef time_t
 #  define time_t time_t
-typedef int time_t;
+typedef long long time_t;
 # endif
 # ifndef timeval
 #  define timeval timeval
